cleaning up ekf

+++ SOC_Particle/src/Chemistry_BMS.cpp
@@ -69,11 +69,12 @@ void Chemistry::assign_all_chm()
     // 20230401:  Hysteresis tuned to soc=0.7 step data
     const uint8_t M_T = 5;    // Number temperature breakpoints for voc table
     const uint8_t N_S = 18;   // Number soc breakpoints for voc table
-    float Y_T[M_T] = // Temperature breakpoints for voc table
+
+    double Y_T[M_T] = // Temperature breakpoints for voc table
         {5., 11.1, 20., 30., 40.};
-    float X_SOC[N_S] = // soc breakpoints for voc table
+    double X_SOC[N_S] = // soc breakpoints for voc table
         {-0.15, 0.00, 0.05, 0.10,  0.14,  0.17,  0.20,  0.25,  0.30,  0.40,  0.50,  0.60,  0.70,  0.80,  0.90,  0.99,  0.995, 1.00};
-    float T_VOC[M_T * N_S] = // r(soc, dv) table  20250611 data shows this original fits data best T=18C
+    double T_VOC[M_T * N_S] = // r(soc, dv) table  20250611 data shows this original fits data best T=18C
         {4.00, 4.00, 4.00,  4.00,  10.10, 11.60, 12.35, 12.60, 12.67, 12.80, 12.81, 12.88, 12.95, 13.01, 13.07, 13.12, 13.49, 14.45,
          4.00, 4.00, 4.00,  9.40,  11.90, 12.40, 12.60, 12.70, 12.80, 12.86, 12.91, 12.96, 13.01, 13.07, 13.10, 13.13, 13.50, 14.46,
          4.00, 4.00, 10.00, 12.50, 12.67, 12.75, 12.79, 12.85, 12.89, 12.93, 12.94, 12.99, 13.04, 13.11, 13.15, 13.17, 13.62, 14.50,
@@ -86,35 +87,35 @@ void Chemistry::assign_all_chm()
         //  4.00, 4.00, 12.00, 12.65, 12.75, 12.80, 12.85, 12.95, 13.00, 13.08, 13.12, 13.16, 13.20, 13.24, 13.26, 13.27, 13.72, 14.50,
         //  4.00, 4.00, 12.00, 12.65, 12.75, 12.80, 12.85, 12.95, 13.00, 13.08, 13.12, 13.16, 13.20, 13.24, 13.26, 13.27, 13.72, 14.50};
         // used before 6/2025
-        // float T_VOC[M_T * N_S] = // r(soc, dv) table  dag 20230726 tune by 0.3 nominal because data during slow discharge at -0.3 hysteresis
+        // double T_VOC[M_T * N_S] = // r(soc, dv) table  dag 20230726 tune by 0.3 nominal because data during slow discharge at -0.3 hysteresis
         //     {4.00, 4.00, 4.00,  4.00,  10.50, 12.00, 12.75, 13.00, 13.07, 13.20, 13.21, 13.28, 13.35, 13.41, 13.47, 13.52, 13.69, 14.25,
         //     4.00, 4.00, 4.00,  9.80,  12.30, 12.80, 13.00, 13.10, 13.20, 13.26, 13.31, 13.36, 13.41, 13.47, 13.50, 13.53, 13.70, 14.26,
         //     4.00, 4.00, 10.30, 12.90, 13.07, 13.15, 13.19, 13.25, 13.29, 13.33, 13.34, 13.39, 13.44, 13.51, 13.55, 13.57, 13.82, 14.30,
         //     4.00, 4.00, 12.30, 12.95, 13.05, 13.10, 13.15, 13.25, 13.30, 13.38, 13.42, 13.46, 13.50, 13.54, 13.56, 13.57, 13.82, 14.30,
         //     4.00, 4.00, 12.30, 12.95, 13.05, 13.10, 13.15, 13.25, 13.30, 13.38, 13.42, 13.46, 13.50, 13.54, 13.56, 13.57, 13.82, 14.30};
     const uint8_t N_N = 5;                                          // Number of temperature breakpoints for x_soc_min table
-    float X_SOC_MIN[N_N] = {5., 11.1, 20., 30., 40.};      // Temperature breakpoints for soc_min table
-    float T_SOC_MIN[N_N] = {0.10, 0.07, 0.25, 0.25, 0.25}; // soc_min(t).  At 40C BMS shuts off at 12V
-    // float T_SOC_MIN[N_N] = {0.10, 0.07, 0.05, 0.00, 0.20}; // soc_min(t).  At 40C BMS shuts off at 12V
+    double X_SOC_MIN[N_N] = {5., 11.1, 20., 30., 40.};      // Temperature breakpoints for soc_min table
+    double T_SOC_MIN[N_N] = {0.10, 0.07, 0.25, 0.25, 0.25}; // soc_min(t).  At 40C BMS shuts off at 12V
+    // double T_SOC_MIN[N_N] = {0.10, 0.07, 0.05, 0.00, 0.20}; // soc_min(t).  At 40C BMS shuts off at 12V
 
     // Battleborn Hysteresis
     const uint8_t M_H = 3;     // Number of soc breakpoints in r(soc, dv) table t_r, t_s
     const uint8_t N_H = 7;     // Number of dv breakpoints in r(dv) table t_r, t_s
-    float X_DV[N_H] = // dv breakpoints for r(soc, dv) table t_r. // DAG 6/13/2022 tune x10 to match data
+    double X_DV[N_H] = // dv breakpoints for r(soc, dv) table t_r. // DAG 6/13/2022 tune x10 to match data
         {-0.7, -0.5, -0.3, 0.0, 0.15, 0.3, 0.7};
-    float Y_SOC[M_H] = // soc breakpoints for r(soc, dv) table t_r, t_s
+    double Y_SOC[M_H] = // soc breakpoints for r(soc, dv) table t_r, t_s
         {0.0, 0.5, 0.7};
-    float T_R[M_H * N_H] = // r(soc, dv) table.    // DAG 9/29/2022 tune to match hist data
+    double T_R[M_H * N_H] = // r(soc, dv) table.    // DAG 9/29/2022 tune to match hist data
         {0.019, 0.015, 0.016, 0.009, 0.011, 0.017, 0.030,
         0.014, 0.014, 0.010, 0.008, 0.010, 0.015, 0.015,
         0.016, 0.016, 0.013, 0.005, 0.007, 0.010, 0.010};
-    float T_S[M_H * N_H] = // r(soc, dv) table. Not used yet for BB
+    double T_S[M_H * N_H] = // r(soc, dv) table. Not used yet for BB
         {1, 1, 1, 1, 1, 1, 1,
         1, 1, 1, 1, 1, 1, 1,
         1, 1, 1, 1, 1, 1, 1};
-    float T_DV_MAX[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
+    double T_DV_MAX[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
         {0.7,  0.3,  0.2};
-    float T_DV_MIN[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
+    double T_DV_MIN[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
         {-0.7, -0.5, -0.3};
 #endif
 
@@ -127,36 +128,36 @@ void Chemistry::assign_all_chm()
     // 2024-04-03:  tune to data
     const uint8_t M_T = 3;    // Number temperature breakpoints for voc table
     const uint8_t N_S = 21;   // Number soc breakpoints for voc table
-    float Y_T[M_T] = // Temperature breakpoints for voc table
+    double Y_T[M_T] = // Temperature breakpoints for voc table
         {5.1, 5.2, 21.5};
-    float X_SOC[N_S] = // soc breakpoints for voc table
+    double X_SOC[N_S] = // soc breakpoints for voc table
         {-0.035,   0.000,   0.050,   0.100,   0.108,   0.120,   0.140,   0.170,   0.200,   0.250,   0.300,   0.340,   0.400,   0.500,   0.600,   0.700,   0.800,   0.900,   0.980,   0.990,   1.000};
-    float T_VOC[M_T * N_S] = // r(soc, dv) table
+    double T_VOC[M_T * N_S] = // r(soc, dv) table
         {
         4.000,   4.000,  4.000,   4.000,    4.000,  4.000,  4.000,   4.000,   4.000,   4.000,   9.000,  11.770,  12.700,  12.950,  13.050,  13.100,  13.226,  13.259,  13.264,  13.460,  14.270,
         4.000,   4.000,  4.000,   4.000,    4.000,  4.000,  4.000,   4.000,   4.000,   4.000,   9.000,  11.770,  12.700,  12.950,  13.050,  13.100,  13.226,  13.259,  13.264,  13.460,  14.270,
         4.000,   4.000,  9.0000,  9.500,   11.260, 11.850, 12.400,  12.650,  12.730,  12.810,  12.920,  12.960,  13.020,  13.060,  13.220,  13.280,  13.284,  13.299,  13.310,  13.486,  14.700
         };
     const uint8_t N_N = 4;                                        // Number of temperature breakpoints for x_soc_min table
-    float X_SOC_MIN[N_N] = {0.000,  11.00,  21.5,  40.000, };  // Temperature breakpoints for soc_min table
-    float T_SOC_MIN[N_N] = {0.31,   0.31,   0.1,   0.1, };  // soc_min(t)
+    double X_SOC_MIN[N_N] = {0.000,  11.00,  21.5,  40.000, };  // Temperature breakpoints for soc_min table
+    double T_SOC_MIN[N_N] = {0.31,   0.31,   0.1,   0.1, };  // soc_min(t)
 #elif CHEM == 2
     // 2024-04-24T14-51-24:  tune to data
     const uint8_t M_T = 3;    // Number temperature breakpoints for voc table
     const uint8_t N_S = 28;   // Number soc breakpoints for voc table
-    float Y_T[M_T] = // Temperature breakpoints for voc table
+    double Y_T[M_T] = // Temperature breakpoints for voc table
         {21.5, 25.0, 35.0, };
-    float X_SOC[N_S] = // soc breakpoints for voc table
+    double X_SOC[N_S] = // soc breakpoints for voc table
         {-0.400, -0.300, -0.230, -0.200, -0.150, -0.130, -0.114, -0.044,  0.000,  0.016,  0.032,  0.055,  0.064,  0.114,  0.134,  0.154,  0.183,  0.214,  0.300,  0.400,  0.500,  0.600,  0.700,  0.800,  0.900,  0.960,  0.980,  1.000, }; 
-    float T_VOC[M_T * N_S] = // soc breakpoints for soc_min table
+    double T_VOC[M_T * N_S] = // soc breakpoints for soc_min table
         {
         4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  8.170, 11.285, 12.114, 12.558, 12.707, 12.875, 13.002, 13.054, 13.201, 13.275, 13.284, 13.299, 13.307, 13.310, 14.700, 
         4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  4.000,  7.947, 11.000, 11.946, 12.252, 12.588, 12.670, 12.797, 12.833, 12.864, 12.908, 12.957, 13.034, 13.081, 13.106, 13.159, 13.234, 13.272, 13.286, 13.300, 13.300, 14.760, 
         4.000,  4.000,  6.686,  8.206, 10.739, 12.045, 12.411, 12.799, 12.866, 12.890, 12.914, 12.949, 12.963, 13.037, 13.052, 13.067, 13.089, 13.112, 13.146, 13.196, 13.284, 13.318, 13.320, 13.320, 13.320, 13.320, 13.320, 14.760, 
         };
     const uint8_t N_N = 3; // Number of temperature breakpoints for x_soc_min table
-    float X_SOC_MIN[N_N] = {21.5,  25.0, 35.0, };  // Temperature breakpoints for soc_min table
-    float T_SOC_MIN[N_N] = {0.13, -0.04, -0.2, };  // soc_min(t)  ****EXTENDED MIN for model bms_off testing
+    double X_SOC_MIN[N_N] = {21.5,  25.0, 35.0, };  // Temperature breakpoints for soc_min table
+    double T_SOC_MIN[N_N] = {0.13, -0.04, -0.2, };  // soc_min(t)  ****EXTENDED MIN for model bms_off testing
 #endif
 
 
@@ -164,23 +165,23 @@ void Chemistry::assign_all_chm()
 #if( CHEM == 1 || CHEM == 2 )
     const uint8_t M_H = 4;     // Number of soc breakpoints in r(soc, dv) table t_r, t_s
     const uint8_t N_H = 10;    // Number of dv breakpoints in r(dv) table t_r, t_s
-    float X_DV[N_H] = // dv breakpoints for r(soc, dv) table t_r, t_s
+    double X_DV[N_H] = // dv breakpoints for r(soc, dv) table t_r, t_s
         {-.10, -.05, -.04, 0.0, .02, .04, .05, .06, .07, .10};
-    float Y_SOC[M_H] = // soc breakpoints for r(soc, dv) table t_r, t_s
+    double Y_SOC[M_H] = // soc breakpoints for r(soc, dv) table t_r, t_s
         {.47, .75, .80, .86};
-    float T_R[M_H * N_H] = // r(soc, dv) table
+    double T_R[M_H * N_H] = // r(soc, dv) table
         {0.003, 0.003, 0.4, 0.4, 0.4, 0.4, 0.010, 0.010, 0.010, 0.010,
         0.004, 0.004, 0.4, 0.4, 0.4, 0.4, 0.4, 0.4, 0.014, 0.012,
         0.004, 0.004, 0.4, 0.4, 0.4, 0.4, 0.4, 0.4, 0.014, 0.012,
         0.004, 0.004, 0.4, 0.4, .2, .09, 0.04, 0.006, 0.006, 0.006};
-    float T_S[M_H * N_H] = // s(soc, dv) table
+    double T_S[M_H * N_H] = // s(soc, dv) table
         {1., 1., .2, .2, .2, .2, 1., 1., 1., 1.,
         1., 1., .2, .2, .2, 1., 1., 1., 1., 1.,
         1., 1., .2, .2, .2, 1., 1., 1., 1., 1.,
         1., 1., .1, .1, .2, 1., 1., 1., 1., 1.};
-    float T_DV_MAX[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
+    double T_DV_MAX[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
         {0.06, 0.1, 0.1, 0.06};
-    float T_DV_MIN[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
+    double T_DV_MIN[M_H] = // dv_max(soc) table.  Pulled values from insp of T_R where flattens
         {-0.06, -0.06, -0.06, -0.06};
 #endif
 
@@ -263,7 +264,7 @@ void Chemistry::assign_CH()
 
 
 // Workhorse assignment function for Hysteresis
-void Chemistry::assign_hys(const int _n_h, const int _m_h, float *x, float *y, float *t, float *s, float *tx, float *tn)
+void Chemistry::assign_hys(const int _n_h, const int _m_h, double *x, double *y, double *t, double *s, double *tx, double *tn)
 {
     hys_T_ = new TableInterp2D(_n_h, _m_h, x, y, t);
     hys_Tn_ = new TableInterp1D(_m_h, y, tn);
@@ -272,13 +273,13 @@ void Chemistry::assign_hys(const int _n_h, const int _m_h, float *x, float *y, f
 }
 
 // Workhorse assignment function for VOC_SOC
-void Chemistry::assign_voc_soc(const int _n_s, const int _m_t, float *x, float *y, float *t)
+void Chemistry::assign_voc_soc(const int _n_s, const int _m_t, double *x, double *y, double *t)
 {
     voc_T_ = new TableInterp2D(_n_s, _m_t, x, y, t);
 }
 
 // Workhorse assignment function for SOC_MIN
-void Chemistry::assign_soc_min(const int _n_n, float *x, float *t)
+void Chemistry::assign_soc_min(const int _n_n, double *x, double *t)
 {
     soc_min_T_ = new TableInterp1D(_n_n, x, t);
 }
@@ -300,7 +301,7 @@ String Chemistry::decode(const uint8_t mod)
 }
 
 // lookup_voc
-float Chemistry::lookup_voc(const float soc, const double tb_f)
+double Chemistry::lookup_voc(const double soc, const double tb_f)
 {
     return voc_T_->interp(soc, tb_f) + dvoc;
 }
