+++ SOC_Particle/src/myLibrary/myTables.h
@@ -1,28 +1,28 @@
 #ifndef _myTables_h
 #define _myTables_h
-// #define t_float float
+// #define t_double double
 
 // Interpolating, clipping, 1 and 2-D arbitrarily spaced table look-up
 
-void binsearch(float x, float *v, int n, int *high, int *low, float *dx);
-float tab1(float x, float *v, float *y, int n);
-float tab1clip(float x, float *v, float *y, int n);
-float tab2(float x1, float x2, float *v1, float *v2, float *y, int n1, int n2);
+void binsearch(double x, double *v, int n, int *high, int *low, double *dx);
+double tab1(double x, double *v, double *y, int n);
+double tab1clip(double x, double *v, double *y, int n);
+double tab2(double x1, double x2, double *v1, double *v2, double *y, int n1, int n2);
 
 class TableInterp
 {
 public:
   TableInterp();
-  TableInterp(const unsigned int n, float x[]);
+  TableInterp(const unsigned int n, double x[]);
   virtual ~TableInterp();
   // operators
   // functions
-  virtual float interp(void);
+  virtual double interp(void);
   void pretty_print();
 protected:
   unsigned int n1_;
-  float *x_;
-  float *v_;
+  double *x_;
+  double *v_;
 };
 
 // 1-D Interpolation Table Lookup
@@ -30,11 +30,11 @@ class TableInterp1D : public TableInterp
 {
 public:
   TableInterp1D();
-  TableInterp1D(const unsigned int n, float x[], float v[]);
+  TableInterp1D(const unsigned int n, double x[], double v[]);
   ~TableInterp1D();
   //operators
   //functions
-  virtual float interp(float x);
+  virtual double interp(double x);
 
 protected:
 };
@@ -44,11 +44,11 @@ class TableInterp1Dclip : public TableInterp
 {
 public:
   TableInterp1Dclip();
-  TableInterp1Dclip(const unsigned int n, float x[], float v[]);
+  TableInterp1Dclip(const unsigned int n, double x[], double v[]);
   ~TableInterp1Dclip();
   //operators
   //functions
-  virtual float interp(float x);
+  virtual double interp(double x);
 
 protected:
 };
@@ -58,24 +58,24 @@ class TableInterp2D : public TableInterp
 {
 public:
   TableInterp2D();
-  TableInterp2D(const unsigned int n, const unsigned int m, float x[],
-                float y[], float v[]);
+  TableInterp2D(const unsigned int n, const unsigned int m, double x[],
+                double y[], double v[]);
   ~TableInterp2D();
   //operators
-  void put_dx(const float inp) { dx_ = inp; }
-  void put_dy(const float inp) { dy_ = inp; }
-  void put_dz(const float inp) { dz_ = inp; }
+  void put_dx(const double inp) { dx_ = inp; }
+  void put_dy(const double inp) { dy_ = inp; }
+  void put_dz(const double inp) { dz_ = inp; }
 
   //functions
-  virtual float interp(float x, float y);
+  virtual double interp(double x, double y);
   void pretty_print();
 
 protected:
-  float dx_;  // Bias on input into table lookup
-  float dy_;  // Bias on input into table lookup
-  float dz_;  // Bias on calculated output of table lookup
+  double dx_;  // Bias on input into table lookup
+  double dy_;  // Bias on input into table lookup
+  double dz_;  // Bias on calculated output of table lookup
   unsigned int n2_;
-  float *y_;
+  double *y_;
 };
 
 #endif
